(PC-6277) Split hooks to get home content
